pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cutmaster-ai"
version = "0.2.0"
description = "Maximum-control MCP server for DaVinci Resolve Studio — 241 tools, AI-enhanced workflows, skills, and agents."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Celavii" },
]
keywords = ["mcp", "davinci-resolve", "video-editing", "automation", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "fastmcp>=3.0.0",
    "pydantic>=2.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
ai = [
    "google-genai>=1.64.0",
]
interchange = [
    "opentimelineio>=0.18.1",
    "otio-fcp-adapter>=1.0.0",
]
panel = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.0",
    "anthropic>=0.40",
]
all = [
    "cutmaster-ai[ai,interchange,panel]",
]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0",
    "pyright>=1.1",
    "pre-commit>=3.0",
]

[project.scripts]
cutmaster-ai = "cutmaster_ai:main"
cutmaster-ai-panel = "cutmaster_ai.http.app:main"

[project.urls]
Homepage = "https://github.com/CelaviiHQ/cutmaster-ai"
Repository = "https://github.com/CelaviiHQ/cutmaster-ai"

[tool.hatch.build.targets.wheel]
packages = ["src/cutmaster_ai"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "E501",    # line too long — handled by formatter
    "B904",    # raise from
    "SIM105",  # contextlib.suppress — try/except/pass is clearer for Resolve API guards
    "SIM108",  # ternary
]

[tool.ruff.lint.per-file-ignores]
"src/cutmaster_ai/__init__.py" = ["E402", "I001"]
"src/cutmaster_ai/resolve.py" = ["E402"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
testpaths = ["tests"]
